requirements.txt
pytesseract 
pillow
opencv-python
easyocr
paddleocr
paddlepaddle>=2.0.0
-f https://www.paddlepaddle.org.cn/whl/linux/mkl/avx/stable.html
requests
google
notebook 
jupyterlab 
ipykernel
matplotlib
pyspellchecker
spacy
spellchecker
imutils
huggingface_hub
tqdm
sqlalchemy 
psycopg2-binary 
pymilvus 
sentence-transformers 
more-itertools
fastapi 
uvicorn 
pymysql
langchain_groq
chainlit
